requirements.txt
django
environ
djangorestframework
djangorestframework-simplejwt
django-cors-headers
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
apscheduler